Use configureCabalFlags from cabal2nix
The function `configureCabalFlags` has a list of packages
that should be called with certain flags when passing them
to `cabal2nix`.

For example, for `hslua`, it specifies that it is to used the "system"
version of lua (the one provided by nix) and not the version of lua
that comes embedded in the package. Notice that cabal2nix assumes
assumes this flag in newer versions of its `postProcess` method, and
if the flag is not given, it will fail attempting to replace package
"lua" by "lua_5_3".